requirements.txt
playwright==1.40.0
langchain==0.1.0
openai==1.3.0
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.0
requests==2.31.0
beautifulsoup4==4.12.0